backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.30.6
pydantic==2.12.0
SQLAlchemy==2.0.44
psycopg[binary,pool]==3.2.1
alembic==1.13.2
celery==5.5.3
redis==5.0.7
httpx==0.28.1
tenacity==9.0.0
aiobreaker==1.1.0
python-jose==3.3.0
argon2-cffi==23.1.0
python-dotenv==1.0.1
email-validator==2.2.0
